requirements.txt
# Core ML/DL dependencies
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=9.5.0

# Data science and visualization
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
scikit-learn>=1.3.0

# MLOps and experiment tracking
mlflow>=2.4.0
wandb>=0.15.0
tensorboard>=2.13.0

# Configuration and utilities
PyYAML>=6.0
hydra-core>=1.3.0
python-dotenv>=1.0.0
tqdm>=4.65.0
rich>=13.0.0

# Model optimization and deployment
onnx>=1.14.0
onnxruntime>=1.15.0
tensorrt>=8.6.0  # Optional, for NVIDIA GPUs

# Testing and code quality
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# API and web services (optional)
fastapi>=0.100.0
uvicorn>=0.22.0
gradio>=3.35.0

# Video processing
imageio>=2.31.0
imageio-ffmpeg>=0.4.8
